TP4 finie

+++ src/produit/IProduit.java
@@ -6,6 +6,6 @@ public interface IProduit {
 
 	String descriptionProduit();
 
-	double calculerPrix(double prix);
+	int calculerPrix(int prix);
 
 }
